Bugfixes for asserting file content blocks

+++ PowerShell/Modules/Files.psm1
@@ -62,6 +62,9 @@ function Assert-FileContentBlock {
         [string] $LineEnding = "`r`n"
     )
 
+    # Ensure proper line endings in new content
+    $Content = Assert-LineEndings -Text $Content -LineEnding $LineEnding
+
     # Load the contents of the file
     Assert-File -Path $Path
     $existingContent = Get-Content -Path $Path -Raw
@@ -98,23 +101,24 @@ function Assert-FileContentBlock {
 
         # Attempt to replace any existing content
         $contentReplaced = $false
-        $newContent = [Regex]::Replace( `
-                $existingContent, `
-                "^(?<pre>.*?)$Find(?<post>.*?)$", `
-                $matchEvaluator, `
-                [RegexOptions]::ExplicitCapture -bor [RegexOptions]::Singleline
+        $newContent = [Regex]::Replace(
+            $existingContent,
+            "^(?<pre>.*?)$Find(?<post>.*?)$",
+            $matchEvaluator,
+            [RegexOptions]::ExplicitCapture -bor [RegexOptions]::Singleline
         )
 
         if ($contentReplaced) {
-            # If the content was replaced, save the updated file contents
-            Write-Host "Replacing old content in '$Path'"
-            Set-Content -Path $Path -Value $newContent
+            if ($newContent -cne $existingContent) {
+                # If the content was actually updated, replace the file contents
+                Write-Host "Replacing old content in '$Path'"
+                Set-Content -Path $Path -Value $newContent
+            }
         }
         else {
-            # If the content wasn't found, append it to the end of the file
+            # If the content wasn't found, just append it to the end of the file
             Write-Host "Appending content to '$Path'"
-            $newContent = "$existingContent$($LineEnding * 2)$Content"
-            Set-Content -Path $Path -Value $newContent
+            Set-Content -Path $Path -Value "$existingContent$($LineEnding * 2)$Content"
         }
     }
 }
